+++ Scripts/Systems/Hybrid/DSC_ADS_DestroyNullGameObject.cs
@@ -4,7 +4,7 @@ using Unity.Entities;
 namespace DSC.Actor.DOTS
 {
     [UpdateInGroup(typeof(InitializationSystemGroup))]
-    public sealed class DSC_ADS_DestroyNullGameObject : ComponentSystem
+    public sealed class DSC_ADS_DestroyNullGameObject : SystemBase
     {
         protected override void OnUpdate()
         {
@@ -15,7 +15,8 @@ namespace DSC.Actor.DOTS
                 {
                     EntityManager.DestroyEntity(hEntity);
                 }
-            });
+            }).WithoutBurst()
+            .Run();
         }
     }
 }